// script/src/data/interfaces/IDevisCommande.ts
export interface IDevisCommande {
  Numéro: string;
  Devis: string;
  "Date de création": string;
  "Adresse de chantier": string;
  Pro: string;
  "Montant HT": string;
  Type: string;
  Partenaire: string;
  Éditer: string;
  Téléphone: string[];
  Email: string[];
  "Lien unique": string;
}